proydjango/requirements.txt
django == 3.2.3
pillow
django-admin-interface
django-crispy-forms
django-rest-framework